requirements.txt
Django== 5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-decouple==3.8
django_filter==24.3
djangorestframework-simplejwt==5.4.0
drf-yasg==1.21.8